pyproject.toml
[tool.poetry]
name = "philippe-ivan-mbarga-bank-account-v2-d17eb7b1-1904-4f5b-b6a0-462e21dea456"
version = "0.1.0"
description = ""
authors = ["masterivanic <masterivanic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^25.4.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.18.2"
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "./src"]
testpaths = ["tests"]
